requirements.txt
pyodbc
pandas
python-dotenv
google-generativeai
beautifulsoup4
requests
newspaper3k
lxml_html_clean
readability-lxml 
lxml